Support of installation in current working directory

+++ bin/install.js
@@ -9,26 +9,60 @@ const dependencies = require('../package.json').devDependencies;
 const scripts = `"start": "gulp",\n\t\t"build": "gulp build"`;
 const gitignore = `node_modules\ndist`;
 
-const projectFolder = process.argv[2];
+const projectName = process.argv[2];
 
 /**
- * Check whether project folder is valid
+ * Get flags from arguments
  *
- * @param {string} name
- * @return {boolean}
+ * @param {Array} args
+ * @return {Array}
  */
-const checkProjectFolder = (name) => name.length > 0;
+const getFlagsFromArguments = args =>
+  args.slice(3, args.length)
+    .filter((val, index) => val.startsWith('--'));
 
 /**
- * Initialize project
+ * Get options from flags
  *
- * @param {string} projectFolder
+ * @param {Array} flags
+ * @return {Object}
+ */
+const getOptionsFromFlags = flags => ({
+  installInCurrentDir: flags.includes('--current-dir'),
+});
+
+/**
+ * Check whether project name is valid
+ *
+ * @param {String} name
+ * @return {Boolean}
+ */
+const checkProjectName = (name) => name !== undefined && name.length > 0;
+
+/**
+ * Get project folder from name
+ *
+ * @param {String} projectName
+ * @return {String}
+ */
+const getProjectFolderFromName = (projectName) => options.installInCurrentDir ? '' : `${ projectName }`;
+
+/**
+ * Initialize project in specified folder
+ *
+ * @param {String} projectFolder
  * @return {Promise<boolean>}
  */
 const init = projectFolder => {
   return new Promise((resolve, reject) => {
+    let initCommand = 'npm init -f';
+
+    if (projectFolder !== '') {
+      initCommand = `mkdir ${ projectFolder } && cd ${ projectFolder } && ` + initCommand.substr(0);
+    }
+
     exec(
-      `mkdir ${ projectFolder } && cd ${ projectFolder } && npm init -f`,
+      initCommand,
       (err, stdout, stderr) => {
         if (err) {
           console.error(
@@ -38,7 +72,7 @@ const init = projectFolder => {
           reject(false);
         }
 
-        const packageJson = `${ projectFolder }/package.json`;
+        const packageJson = (projectFolder !== '') ? `${ projectFolder }/package.json` : 'package.json';
 
         fs.readFile(packageJson, (readErr, file) => {
           if (readErr) throw readErr;
@@ -50,7 +84,11 @@ const init = projectFolder => {
           fs.writeFile(packageJson, data, writeErr => writeErr || true);
         });
 
-        const filesToCopy = [ 'gulpfile.js', '.editorconfig' ];
+        const filesToCopy = [
+          'gulpfile.js',
+          '.editorconfig'
+        ].filter((file, index) =>
+          !fs.existsSync(path.join(__dirname, `../${ file }`)));
 
         for (let i = 0; i < filesToCopy.length; i++) {
           fs
@@ -58,7 +96,9 @@ const init = projectFolder => {
             .pipe(fs.createWriteStream(`${ projectFolder }/${ filesToCopy[i] }`));
         }
 
-        fs.writeFile(`${ projectFolder }/.gitignore`, gitignore, writeErr => writeErr || true);
+        if (!fs.existsSync(`${ projectFolder }/.gitignore`)) {
+          fs.writeFile(`${projectFolder}/.gitignore`, gitignore, writeErr => writeErr || true);
+        }
 
         resolve(true);
       }
@@ -69,8 +109,8 @@ const init = projectFolder => {
 /**
  * Get a formatted string of all dependencies to install
  *
- * @param {object} dependencies
- * @return {string}
+ * @param {Object} dependencies
+ * @return {String}
  */
 const getFormattedDependencies = dependencies =>
   Object.entries(dependencies)
@@ -82,15 +122,16 @@ const getFormattedDependencies = dependencies =>
 /**
  * Install dependencies
  *
- * @param {object} dependencies
+ * @param {Object} dependencies
+ * @param {String} directory
  * @return {Promise<boolean>}
  */
-const installDependencies = dependencies => {
+const installDependencies = (dependencies, directory) => {
   return new Promise((resolve, reject) => {
     dependencies = getFormattedDependencies(dependencies);
 
     exec(
-      `cd ${ projectFolder } && npm i --save-dev ${ dependencies }`,
+      `cd ${ directory } && npm i --save-dev ${ dependencies }`,
       (npmErr, npmStdout) => {
         if (npmErr) {
           console.error(
@@ -109,14 +150,15 @@ const installDependencies = dependencies => {
 };
 
 /**
- * Copy additional files
+ * Copy additional files into directory
  *
- * @return {boolean}
+ * @param {String} directory
+ * @return {Boolean}
  */
-const copyAdditionalFiles = async () => {
+const copyAdditionalFiles = async directory => {
   try {
     await fs
-      .copy(path.join(__dirname, '../src'), `${projectFolder}/src`);
+      .copy(path.join(__dirname, '../src'), `${ directory }/src`);
     return true;
   } catch (err) {
     console.error(err);
@@ -125,39 +167,50 @@ const copyAdditionalFiles = async () => {
 };
 
 /**
- * Install gulp-starter-kit into project folder
+ * Install gulp-starter-kit project
  *
- * @param {string} projectFolder
- * @param {object} dependencies
- * @return {Promise<boolean>}
+ * @param {String} projectName
+ * @param {Object} dependencies
+ * @param {Object} options
+ * @return {Promise<boolean>,Promise<string>}
  */
-const install = async (projectFolder, dependencies) => {
-  console.log('npm init - Initializing your project...');
+const install = async (projectName, dependencies, options) => {
+  return new Promise(async (resolve, reject) => {
+    console.log('npm init - Initializing your project...');
 
-  if (! await init(projectFolder)) return false;
+    const projectFolder = getProjectFolderFromName(projectName);
 
-  console.log('npm init successful - Your npm package has been initialized');
+    if (! await init(projectFolder)) reject(false);
 
-  console.log('Installing dependencies - This might take a few minutes...');
+    console.log('npm init successful - Your npm package has been initialized');
 
-  if (! await installDependencies(dependencies)) return false;
+    console.log('Installing dependencies - This might take a few minutes...');
 
-  console.log('Dependency installation successful - All dependencies have been installed');
+    if (! await installDependencies(dependencies, projectFolder)) reject(false);
 
-  console.log('Copying additional files...');
+    console.log('Dependency installation successful - All dependencies have been installed');
 
-  if (!copyAdditionalFiles()) return false;
+    console.log('Copying additional files...');
 
-  console.log('Copying additional files successful');
+    if (!copyAdditionalFiles(projectFolder)) reject(false);
 
-  return true;
+    console.log('Copying additional files successful');
+
+    resolve(projectFolder);
+  });
 };
 
-if (!checkProjectFolder(projectFolder)) {
+const options = getOptionsFromFlags(
+  getFlagsFromArguments(process.argv)
+);
+
+if (!checkProjectName(projectName)) {
   console.error('Oops, looks like you have not specified any project name. Please make sure to do that.');
   return;
 }
 
-install(projectFolder, dependencies).then(() => {
+install(projectName, dependencies, options).then((projectFolder) => {
   console.log(`\nAll done!\nYour project has been set up to the ${ projectFolder } folder.\nHappy Coding!`);
+}).catch(() => {
+  console.error('Oops, looks like something went wrong installing the Gulp Starter Kit.');
 });
